requirements.txt
python-binance
ta
rich